CubicFunctionをdoubleにする

+++ Assets/Editor/Test/CubicFunction.cs
@@ -13,8 +13,8 @@ namespace TSKT.Tests
         public void Solve2Points(float t1, float v1, float t2, float v2)
         {
             var f = TSKT.CubicFunction.Solve2Points((t1, v1), (t2, v2));
-            Assert.True(Mathf.Approximately(v1, f.Evaluate(t1)), v1 + ", " + f.Evaluate(t1));
-            Assert.True(Mathf.Approximately(v2, f.Evaluate(t2)), v2 + ", " + f.Evaluate(t2));
+            Assert.True(Mathf.Approximately(v1, (float)f.Evaluate(t1)), v1 + ", " + f.Evaluate(t1));
+            Assert.True(Mathf.Approximately(v2, (float)f.Evaluate(t2)), v2 + ", " + f.Evaluate(t2));
             Assert.AreEqual(0f, f.a);
             Assert.AreEqual(0f, f.b);
         }
@@ -24,10 +24,10 @@ namespace TSKT.Tests
         public void Solve2PointsAnd2Velocities(float t1, float v1, float t2, float v2, float t3, float v3, float t4, float v4)
         {
             var f = TSKT.CubicFunction.Solve2PointsAnd2Velocities((t1, v1), (t2, v2), (t3, v3), (t4, v4));
-            Assert.True(Mathf.Approximately(v1, f.Evaluate(t1)));
-            Assert.True(Mathf.Approximately(v2, f.Evaluate(t2)));
-            Assert.True(Mathf.Approximately(v3, f.Velocity(t3)));
-            Assert.True(Mathf.Approximately(v4, f.Velocity(t4)));
+            Assert.True(Mathf.Approximately(v1, (float)f.Evaluate(t1)));
+            Assert.True(Mathf.Approximately(v2, (float)f.Evaluate(t2)));
+            Assert.True(Mathf.Approximately(v3, (float)f.Velocity(t3)));
+            Assert.True(Mathf.Approximately(v4, (float)f.Velocity(t4)));
         }
 
         [Test]
@@ -35,9 +35,9 @@ namespace TSKT.Tests
         public void Solve2PointsAndConstantAccel(float t1, float v1, float t2, float v2, float a)
         {
             var f = TSKT.CubicFunction.Solve2PointsAndConstantAccel((t1, v1), (t2, v2), a);
-            Assert.True(Mathf.Approximately(v1, f.Evaluate(t1)), v1 + ", " + f.Evaluate(t1));
-            Assert.True(Mathf.Approximately(v2, f.Evaluate(t2)));
-            Assert.True(Mathf.Approximately(a, f.Acceleration(0f)));
+            Assert.True(Mathf.Approximately(v1, (float)f.Evaluate(t1)), v1 + ", " + f.Evaluate(t1));
+            Assert.True(Mathf.Approximately(v2, (float)f.Evaluate(t2)));
+            Assert.True(Mathf.Approximately(a, (float)f.Acceleration(0f)));
             Assert.AreEqual(0f, f.a);
         }
 
@@ -46,9 +46,9 @@ namespace TSKT.Tests
         public void Solve2PointsAnd2Velocities(float t1, float v1, float t2, float v2, float t3, float v3)
         {
             var f = TSKT.CubicFunction.Solve2PointsAndVelocity((t1, v1), (t2, v2), (t3, v3));
-            Assert.True(Mathf.Approximately(v1, f.Evaluate(t1)));
-            Assert.True(Mathf.Approximately(v2, f.Evaluate(t2)));
-            Assert.True(Mathf.Approximately(v3, f.Velocity(t3)));
+            Assert.True(Mathf.Approximately(v1, (float)f.Evaluate(t1)));
+            Assert.True(Mathf.Approximately(v2, (float)f.Evaluate(t2)));
+            Assert.True(Mathf.Approximately(v3, (float)f.Velocity(t3)));
             Assert.AreEqual(0f, f.a);
         }
 
@@ -57,9 +57,9 @@ namespace TSKT.Tests
         public void SolvePointAndVelocityAndConstantAccel(float t1, float p1, float t2, float v2, float a)
         {
             var f = TSKT.CubicFunction.SolvePointAndVelocityAndConstantAccel((t1, p1), (t2, v2), a);
-            Assert.True(Mathf.Approximately(p1, f.Evaluate(t1)));
-            Assert.True(Mathf.Approximately(v2, f.Velocity(t2)));
-            Assert.True(Mathf.Approximately(a, f.Acceleration(0f)));
+            Assert.True(Mathf.Approximately(p1, (float)f.Evaluate(t1)));
+            Assert.True(Mathf.Approximately(v2, (float)f.Velocity(t2)));
+            Assert.True(Mathf.Approximately(a, (float)f.Acceleration(0f)));
             Assert.AreEqual(0f, f.a);
         }
 
@@ -68,10 +68,10 @@ namespace TSKT.Tests
         public void Solve2PointsVelocityAndAccel(float t1, float v1, float t2, float v2, float t3, float v3, float t4, float v4)
         {
             var f = TSKT.CubicFunction.Solve2PointsVelocityAndAccel((t1, v1), (t2, v2), (t3, v3), (t4, v4));
-            Assert.True(Mathf.Approximately(v1, f.Evaluate(t1)), v1 + ", " + f.Evaluate(t1));
-            Assert.True(Mathf.Approximately(v2, f.Evaluate(t2)), v2 + ", " + f.Evaluate(t2));
-            Assert.True(Mathf.Approximately(v3, f.Velocity(t3)), v3 + ", " + f.Velocity(t3));
-            Assert.True(Mathf.Approximately(v4, f.Acceleration(t4)));
+            Assert.True(Mathf.Approximately(v1, (float)f.Evaluate(t1)), v1 + ", " + f.Evaluate(t1));
+            Assert.True(Mathf.Approximately(v2, (float)f.Evaluate(t2)), v2 + ", " + f.Evaluate(t2));
+            Assert.True(Mathf.Approximately(v3, (float)f.Velocity(t3)), v3 + ", " + f.Velocity(t3));
+            Assert.True(Mathf.Approximately(v4, (float)f.Acceleration(t4)));
         }
 
         [Test]
@@ -79,9 +79,9 @@ namespace TSKT.Tests
         public void Solve3Points(float t1, float v1, float t2, float v2, float t3, float v3)
         {
             var f = TSKT.CubicFunction.Solve3Points((t1, v1), (t2, v2), (t3, v3));
-            Assert.True(Mathf.Approximately(v1, f.Evaluate(t1)), v1 + ", " + f.Evaluate(t1));
-            Assert.True(Mathf.Approximately(v2, f.Evaluate(t2)), v2 + ", " + f.Evaluate(t2));
-            Assert.True(Mathf.Approximately(v3, f.Evaluate(t3)), v2 + ", " + f.Evaluate(t3));
+            Assert.True(Mathf.Approximately(v1, (float)f.Evaluate(t1)), v1 + ", " + f.Evaluate(t1));
+            Assert.True(Mathf.Approximately(v2, (float)f.Evaluate(t2)), v2 + ", " + f.Evaluate(t2));
+            Assert.True(Mathf.Approximately(v3, (float)f.Evaluate(t3)), v2 + ", " + f.Evaluate(t3));
             Assert.AreEqual(0f, f.a);
         }
 
@@ -90,10 +90,10 @@ namespace TSKT.Tests
         public void Solve3PointsAndVelocity(float t1, float v1, float t2, float v2, float t3, float v3, float t4, float v4)
         {
             var f = TSKT.CubicFunction.Solve3PointsAndVelocity((t1, v1), (t2, v2), (t3, v3), (t4, v4));
-            Assert.True(Mathf.Approximately(v1, f.Evaluate(t1)), v1 + ", " + f.Evaluate(t1));
-            Assert.True(Mathf.Approximately(v2, f.Evaluate(t2)), v2 + ", " + f.Evaluate(t2));
-            Assert.True(Mathf.Approximately(v3, f.Evaluate(t3)), v3 + ", " + f.Evaluate(t3));
-            Assert.True(Mathf.Approximately(v4, f.Velocity(t4)));
+            Assert.True(Mathf.Approximately(v1, (float)f.Evaluate(t1)), v1 + ", " + f.Evaluate(t1));
+            Assert.True(Mathf.Approximately(v2, (float)f.Evaluate(t2)), v2 + ", " + f.Evaluate(t2));
+            Assert.True(Mathf.Approximately(v3, (float)f.Evaluate(t3)), v3 + ", " + f.Evaluate(t3));
+            Assert.True(Mathf.Approximately(v4, (float)f.Velocity(t4)));
         }
 
         [Test]
@@ -101,10 +101,10 @@ namespace TSKT.Tests
         public void Solve4Points(float t1, float v1, float t2, float v2, float t3, float v3, float t4, float v4)
         {
             var f = TSKT.CubicFunction.Solve4Points((t1, v1), (t2, v2), (t3, v3), (t4, v4));
-            Assert.True(Mathf.Approximately(v1, f.Evaluate(t1)), v1 + ", " + f.Evaluate(t1));
-            Assert.True(Mathf.Approximately(v2, f.Evaluate(t2)), v2 + ", " + f.Evaluate(t2));
-            Assert.True(Mathf.Approximately(v3, f.Evaluate(t3)), v3 + ", " + f.Evaluate(t3));
-            Assert.True(Mathf.Approximately(v4, f.Evaluate(t4)), v3 + ", " + f.Evaluate(t4));
+            Assert.True(Mathf.Approximately(v1, (float)f.Evaluate(t1)), v1 + ", " + f.Evaluate(t1));
+            Assert.True(Mathf.Approximately(v2, (float)f.Evaluate(t2)), v2 + ", " + f.Evaluate(t2));
+            Assert.True(Mathf.Approximately(v3, (float)f.Evaluate(t3)), v3 + ", " + f.Evaluate(t3));
+            Assert.True(Mathf.Approximately(v4, (float)f.Evaluate(t4)), v3 + ", " + f.Evaluate(t4));
         }
     }
 }
